Mybatis implementation

+++ src/main/java/com/solvd/delivery/service/impl/CustomerService.java
@@ -7,39 +7,50 @@ import com.solvd.delivery.service.ICustomerService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.List;
+
 
 public class CustomerService implements ICustomerService {
 
     private static final Logger LOGGER = LogManager.getLogger(CustomerService.class);
-
+    private static final ICustomerDAO dao = new CustomerDAO();
 
     @Override
-    public void saveCustomerToDB(Customer customer) {
-        if (customer.getName() != null &&
-            customer.getAddress() != null &&
-            customer.getEmail() != null) {
-            ICustomerDAO dao = new CustomerDAO();
-            dao.insert(customer);
-        }
+    public Customer getCustomerByID(int id) {
+        if (id > 0) {
+            return dao.getByID(id);
+        } else LOGGER.warn("Invalid ID provided! ");
+        return null;
     }
 
     @Override
-    public void showAllCustomers() {
-        ICustomerDAO dao = new CustomerDAO();
-        dao.getAll().forEach(customer -> LOGGER.info(customer.toString()));
+    public void addCustomerToDB(Customer customer) {
+        if (customer.getName() != null &&
+            customer.getAddress() != null &&
+            customer.getEmail() != null) {
+                dao.insert(customer);
+                LOGGER.info("Successfully added to database\n");
+        } else
+            LOGGER.warn("Error inserting customer to database");
+    }
+
+    @Override
+    public List<Customer> getAllCustomers() {
+        return dao.getAll();
     }
 
     @Override
     public void removeCustomerFromDB(Customer customer) {
-        ICustomerDAO dao = new CustomerDAO();
-        dao.deleteByID(customer.getId());
+        if (customer.getId() > 0) {
+            dao.deleteByID(customer.getId());
+        } else LOGGER.warn("Invalid Customer ID provided! ");
     }
 
     @Override
     public void removeCustomerFromDB(int id) {
-        ICustomerDAO dao = new CustomerDAO();
-        dao.deleteByID(id);
+        if (id > 0) {
+            dao.deleteByID(id);
+        } else LOGGER.warn("Invalid ID provided! ");
     }
-
 
 }
